+++ src/Extractor/OptionsExtractor.php
@@ -12,8 +12,8 @@ namespace ZE\ContentValidation\Extractor;
 
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\ServerRequestInterface as Request;
-use Zend\Expressive\Router\RouteResult;
-use Zend\Expressive\Router\RouterInterface;
+use Mezzio\Router\RouteResult;
+use Mezzio\Router\RouterInterface;
 
 /**
  * Class OptionsExtractor