+++ CHANGELOG.md
@@ -6,3 +6,8 @@
 
   * Vagrantfile for easy local testing
   * Licensing information
+
+### Changed
+
+  * Support PHP7.4 only
+  * Migrate to Mezzio/Laminas